Created the graph module and methods for traversal
The shortest path is discovered by using breadth first search.

First, a method is created to have a 2D array variable of the board,
the 2D array will store the coordinates of the previous move of a node.
This method will start from the start node and dynamically explore
the next possible moves, which allows it to traverse to the neighbouring
nodes. As it conducts a BFS, the neighbouring nodes are pushed into a
queue so as long as the end node is found in any of the next possible
moves/ neighbouring nodes, the search is completed and it will return
the 2D array.

A method is used to reconstruct the path based on the 2D array by simply
backtracking what the previous nodes are by traversing from the end node.